Add parser for name of musical forms
As a first step towards listing form names from legends
